+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.3.0] - 2026-05-08
+
 - Replace state atom with separate `@in_flight` and `@shutdown` atoms
 - Add `on_error:` worker error callback
 - Update `Ractor` warning suppression regex